Update ExcelEconomaExport.bas
Adjusted the code copying rows from the source budget sheet to the output budget sheet with a couple of added variables to make the intent of the code clearer. This solves Issue #29
Changes also address issue #31 to make sure the last row of the source sheet isn't included as that row contains totals for the entire budget sheet.